Move `RenderQuest()` method
