+++ nekoyume/Assets/_Scripts/BlockChain/ActionHandler.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using Cysharp.Threading.Tasks;
 using Lib9c.Renderer;
@@ -331,9 +332,9 @@ namespace Nekoyume.BlockChain
         }
 
         internal static void UpdateCombinationSlotState(
-        Address avatarAddress,
-        int slotIndex,
-        CombinationSlotState state)
+            Address avatarAddress,
+            int slotIndex,
+            CombinationSlotState state)
         {
             States.Instance.UpdateCombinationSlotState(avatarAddress, slotIndex, state);
             UpdateCache(state);
@@ -349,6 +350,35 @@ namespace Nekoyume.BlockChain
             if (Game.Game.instance.CachedStates.ContainsKey(state.address))
             {
                 Game.Game.instance.CachedStates[state.address] = state.Serialize();
+            }
+        }
+
+        protected static void RenderQuest(Address avatarAddress, IEnumerable<int> ids)
+        {
+            if (avatarAddress != States.Instance.CurrentAvatarState.address)
+            {
+                return;
+            }
+
+            var questList = States.Instance.CurrentAvatarState.questList;
+            foreach (var id in ids)
+            {
+                var quest = questList.First(q => q.Id == id);
+                var rewardMap = quest.Reward.ItemMap;
+
+                foreach (var reward in rewardMap)
+                {
+                    var materialRow = TableSheets.Instance
+                        .MaterialItemSheet
+                        .First(pair => pair.Key == reward.Key);
+
+                    LocalLayerModifier.RemoveItem(
+                        avatarAddress,
+                        materialRow.Value.ItemId,
+                        reward.Value);
+                }
+
+                LocalLayerModifier.AddReceivableQuest(avatarAddress, id);
             }
         }
     }
